abstract crud page
