+++ hackerforms/crud/postgres/postgres_row_page.py
@@ -5,9 +5,8 @@ from hackerforms.crud.postgres.sql.queries import insert_new_row_query
 
 
 class PostgresRowPage(RowPage):
-    def __init__(self, table, insert_method):
-        super().__init__(table, insert_method)
+    def __init__(self, table, connector):
+        super().__init__(table, connector)
 
-    def insert_data(self, table: str, data: Dict[str, any]):
-        query = insert_new_row_query(table, data)
-        self.connector.insert(query)
+    def get_insert_data_query(self, table: str, data: Dict[str, any]):
+        return insert_new_row_query(table, data)
