+++ hackerforms/crud/row_page.py
@@ -1,146 +1,62 @@
+"""
+hackerforms.crud.row_page
+~~~~~~~~~~~~~~~~~~~~~~~
+
+This module provides a base class for all the row page implementations.
+
+"""
+
+
 from typing import Dict
 
-from hackerforms.generated.page import Page
+from hackerforms.crud.crud_page import CRUDPage
 
+class RowPage(CRUDPage):
+    """Base class for the concrete row page implementations
 
-class RowPage:
+    Its constructor receives the target table (or entity in api)
+    and the the connector to manipulate data source 
+
+    TODO::
+        - update of a table's row value
+    
+    """
     def __init__(self, table, connector):
-        self.table = table
-        self.page = Page()
-        self.column_from_key = {}
         self.connector = connector
-
-    def __validate_args(self, *args, **kwargs):
-        column = kwargs.get("column", None)
-        if not column or not args:
-            raise Exception
-
-    def __add_to_column_map(self, *args, **kwargs):
-        key = kwargs.get("key", None) or args[0]
-        column = kwargs.get("column")
-        self.column_from_key[key] = column
-
-    def read(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read(*args, **kwargs)
-        return self
-
-    def read_code(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_code(*args, **kwargs)
-        return self
-
-    def read_currency(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_currency(*args, **kwargs)
-        return self
-
-    def read_date(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_date(*args, **kwargs)
-        return self
-
-    def read_dropdown(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_dropdown(*args, **kwargs)
-        return self
-
-    def read_email(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_email(*args, **kwargs)
-        return self
-
-    def read_file(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_file(*args, **kwargs)
-        return self
-
-    def read_html_list(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_html_list(*args, **kwargs)
-        return self
-
-    def read_image(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_image(*args, **kwargs)
-        return self
-
-    def read_list(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_list(*args, **kwargs)
-        return self
-
-    def read_multiple_choice(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_multiple_choice(*args, **kwargs)
-        return self
-
-    def read_nps(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_nps(*args, **kwargs)
-        return self
-
-    def read_number(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_number(*args, **kwargs)
-        return self
-
-    def read_pandas_row_selection(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_pandas_row_selection(*args, **kwargs)
-        return self
-
-    def read_password(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_password(*args, **kwargs)
-        return self
-
-    def read_phone(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_phone(*args, **kwargs)
-        return self
-
-    def read_tag(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_tag(*args, **kwargs)
-        return self
-
-    def read_textarea(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_textarea(*args, **kwargs)
-        return self
-
-    def read_time(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_time(*args, **kwargs)
-        return self
-
-    def read_video(self, *args, **kwargs):
-        self.__validate_args(*args, **kwargs)
-        self.__add_to_column_map(*args, **kwargs)
-        self.page = self.page.read_video(*args, **kwargs)
-        return self
+        self.table = table
 
     def insert(self, **kwargs):
+        """It returns the page instance with the form for user submitting
+        a new row value for inserting in the table
+        
+        :param `context` kwargs: a dict with values that must be inserted within the row. Its keys are column names and the values are the respective data to be inserted. Be aware that all columns set here override columns if the same are set mapping to the widgets form.
+        
+         Example::
+
+                >>> categories = db.dropdown(table="categories", column="category_name", primary_key='category_id') # [{'label': 'category_name', 'value': 'category_id'}]
+                >>> suppliers = db.dropdown(table="suppliers", column="company_name", primary_key='supplier_id')
+                >>> page = (
+                ...    db.RowPage(table="products")
+                ...     .read_dropdown("category", options=categories, column="category_id")
+                ...     .read_dropdown("supplier", options=suppliers, column="supplier_id")
+                ...     .read("product name", key="product_name", column="product_name")
+                ...     .read_number("quantity per unit", column="quantity_per_unit")
+                ...     .read_currency("unit price", column="unit_price")
+                ...     .read_number("units in stock", column="units_in_stock")
+                ...     .read_number("units in order", column="units_on_order")
+                ...     .read_number("discontinued", column="discontinued")
+                ...     .insert(context={"product_id": 811, })
+                ... )
+
+            Note: primary keys generations are not currently handled by hackerforms package, so
+            it must be generated and provided in the context in case of its value not be handled
+            by the data source.
+
+            TODO::
+                - deal primary key colision
+                - abstract primary key generation as option
+        """
+
         context = kwargs.get("context", None)
 
         page = self.page.run()
@@ -149,8 +65,10 @@ class RowPage:
         if context:
             data = {**data, **context}
 
-        self.insert_data(self.table, data)
+        insert_query = self.get_insert_data_query(self.table, data)
+        self.connector.insert(insert_query)
         return page
 
-    def insert_data(self, table: str, data: Dict[str, any]):
+    def get_insert_data_query(self, table: str, data: Dict[str, any]):
+        """Abstract method that returns insert data query"""
         pass
